+++ src/lib/logfluent.ts
@@ -27,7 +27,11 @@ class LogFluent implements FluentConfig {
   }
 
   output(hostname: string, logs: LogLine[], printnots: boolean, debug: boolean) {
-    logger.log(logger.debug, `Sending ${logs.length} new log records to FluentD this interval.`, debug);
+    logger.log(logger.debug, `Sending ${logs.length} new log records to FluentD this interval.`, debug, {
+      log_file: logs[0].logfile,
+      log_length: logs.length,
+      hostname: hostname
+    });
 
     for (let j = 0; j < logs.length; j++) {
       let log = logs[j];